feat(Container): 전역 컨테이너 컴포넌트 퍼블리싱 및 분류컴포넌트 병합
